refactor: extract structured-output logic and harden guard placement
